Migrate Select from /docs/components/select to /components/select
Move all examples (Basic, Form, Grouped) from old doc page to the new
reference page and remove the old page. Update navigation links in
routes, mobile menu, command palette, and E2E tests.
