+++ site/ui/components/command-palette.tsx
@@ -30,7 +30,7 @@ const componentItems = [
   { id: 'dialog', title: 'Dialog', href: '/docs/components/dialog', category: 'Components' },
   { id: 'dropdown-menu', title: 'Dropdown Menu', href: '/docs/components/dropdown-menu', category: 'Components' },
   { id: 'input', title: 'Input', href: '/components/input', category: 'Components' },
-  { id: 'select', title: 'Select', href: '/docs/components/select', category: 'Components' },
+  { id: 'select', title: 'Select', href: '/components/select', category: 'Components' },
   { id: 'switch', title: 'Switch', href: '/docs/components/switch', category: 'Components' },
   { id: 'tabs', title: 'Tabs', href: '/docs/components/tabs', category: 'Components' },
   { id: 'toast', title: 'Toast', href: '/docs/components/toast', category: 'Components' },
