+++ site/ui/e2e/home.spec.ts
@@ -23,7 +23,7 @@ test.describe('Home Page', () => {
     await expect(page.locator('#components a[href="/components/card"]')).toBeVisible()
     await expect(page.locator('#components a[href="/docs/components/command"]')).toBeVisible()
     await expect(page.locator('#components a[href="/docs/components/dialog"]')).toBeVisible()
-    await expect(page.locator('#components a[href="/docs/components/select"]')).toBeVisible()
+    await expect(page.locator('#components a[href="/components/select"]')).toBeVisible()
     await expect(page.locator('#components a[href="/docs/components/slider"]')).toBeVisible()
     await expect(page.locator('#components a[href="/docs/components/switch"]')).toBeVisible()
     await expect(page.locator('#components a[href="/docs/components/tabs"]')).toBeVisible()
